# dist/lib/main/json/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Hyperledger Cactus Plugin - Connector Corda",
        "description": "Can perform basic tasks on a Corda ledger",
        "version": "v2.0.0-alpha.2",
        "license": {
            "name": "Apache-2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "components": {
        "schemas": {
...
            "DeployContractJarsV1Request": {
                "type": "object",
                "required": ["jarFiles", "cordappDeploymentConfigs"],
                "additionalProperties": false,
                "properties": {
                    "cordappDeploymentConfigs": {
                        "type": "array",
                        "description": "The list of deployment configurations pointing to the nodes where the provided cordapp jar files are to be deployed .",
                        "minLength": 0,
                        "maxLength": 1024,
                        "default": [],
                        "items": {
                            "$ref": "#/components/schemas/CordappDeploymentConfig"
                        }
                    },
                    "jarFiles": {
                        "type": "array",
                        "nullable": false,
                        "items": {
                            "minLength": 1,
                            "maxLength": 1024,
                            "$ref": "#/components/schemas/JarFile"
                        }
                    }
                }
            },
            "DeployContractJarsSuccessV1Response": {
                "type": "object",
                "required": ["deployedJarFiles"],
                "properties": {
                    "deployedJarFiles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minItems": 1,
                            "nullable": false
                        }
                    }
                }
            },
            "DeployContractJarsBadRequestV1Response": {
                "type": "object",
                "required": ["errors"],
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 65535,
                            "minItems": 1,
                            "maxItems": 2048
                        }
                    }
                }
            },
            "InvokeContractV1Request": {
                "type": "object",
                "required": [
                    "flowFullClassName",
                    "flowInvocationType",
                    "params",
                    "signingCredential"
                ],
                "additionalProperties": false,
                "properties": {
                    "flowFullClassName": {
                        "description": "The fully qualified name of the Corda flow to invoke",
                        "example": "net.corda.samples.obligation.flows.IOUIssueFlow",
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1024,
                        "nullable": false
                    },
                    "flowInvocationType": {
                        "$ref": "#/components/schemas/FlowInvocationType"
                    },
                    "params": {
                        "description": "The list of arguments to pass in to the contract method being invoked.",
                        "type": "array",
                        "default": [],
                        "items": {
                            "$ref": "#/components/schemas/JvmObject",
                            "nullable": false
                        },
                        "nullable": false
                    },
                    "timeoutMs": {
                        "type": "integer",
                        "description": "The amount of milliseconds to wait for a transaction receipt beforegiving up and crashing.",
                        "minimum": 0,
                        "default": 60000,
                        "nullable": false
                    }
                }
            },
            "InvokeContractV1Response": {
                "type": "object",
                "required": ["success", "callOutput", "flowId"],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "nullable": false
                    },
                    "callOutput": {
                        "type": "object",
                        "description": "Data returned from the JVM when no transaction is running"
                    },
                    "transactionId": {
                        "type": "string",
                        "description": "The net.corda.core.flows.StateMachineRunId value returned by the flow execution.",
                        "minLength": 1,
                        "maxLength": 1024,
                        "nullable": false
                    },
                    "progress": {
                        "type": "array",
                        "description": "An array of strings representing the aggregated stream of progress updates provided by a *tracked* flow invocation. If the flow invocation was not tracked, this array is still returned, but as empty.",
                        "items": {
                            "type": "string",
                            "minItems": 0,
                            "maxItems": 10000000
                        },
                        "default": []
                    },
                    "flowId": {
                        "type": "string",
                        "description": "The id for the flow handle"
                    }
                }
            },
            "StartMonitorV1Request": {
                "type": "object",
                "required": ["clientAppId", "stateFullClassName"],
                "additionalProperties": false,
                "properties": {
                    "clientAppId": {
                        "description": "ID of a client application that wants to monitor the state changes",
                        "example": "frond_v1_x_7Hdg6s",
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1024,
                        "nullable": false
                    },
                    "stateFullClassName": {
                        "description": "The fully qualified name of the Corda state to monitor",
                        "example": "net.corda.samples.example.states.IOUState",
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 1024,
                        "nullable": false
                    }
                }
            },
...
            "NetworkMapV1Request": {
                "type": "object"
            },
...
        }
    },
    "paths": {
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/deploy-contract-jars": {
            "post": {
                "operationId": "deployContractJarsV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/deploy-contract-jars",
                        "verbLowerCase": "post"
                    }
                },
                "summary": "Deploys a set of jar files (Cordapps, e.g. the contracts in Corda speak).",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeployContractJarsV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeployContractJarsSuccessV1Response"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeployContractJarsBadRequestV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/invoke-contract": {
            "post": {
                "operationId": "invokeContractV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/invoke-contract",
                        "verbLowerCase": "post"
                    }
                },
                "summary": "Invokes a contract on a Corda ledger (e.g. a flow)",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvokeContractV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvokeContractV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/start-monitor": {
            "post": {
                "operationId": "startMonitorV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/start-monitor",
                        "verbLowerCase": "post"
                    }
                },
                "summary": "Start monitoring corda changes (transactions) of given state class",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartMonitorV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StartMonitorV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/get-monitor-transactions": {
            "get": {
                "operationId": "GetMonitorTransactionsV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/get-monitor-transactions",
                        "verbLowerCase": "get"
                    }
                },
                "summary": "Get transactions for monitored state classes.",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetMonitorTransactionsV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetMonitorTransactionsV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/clear-monitor-transactions": {
            "delete": {
                "operationId": "ClearMonitorTransactionsV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/clear-monitor-transactions",
                        "verbLowerCase": "delete"
                    }
                },
                "summary": "Clear transactions from internal store so they'll not be available by GetMonitorTransactionsV1 anymore.",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClearMonitorTransactionsV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClearMonitorTransactionsV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/stop-monitor": {
            "delete": {
                "operationId": "stopMonitorV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/stop-monitor",
                        "verbLowerCase": "delete"
                    }
                },
                "summary": "Stop monitoring corda changes (transactions) of given state class",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StopMonitorV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StopMonitorV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/network-map": {
            "post": {
                "operationId": "networkMapV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/network-map",
                        "verbLowerCase": "post"
                    }
                },
                "description": "Responds with a snapshot of the network map as provided by the Corda RPC call: net.corda.core.messaging.CordaRPCOps public abstract fun networkMapSnapshot(): List<NodeInfo>",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NetworkMapV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NetworkMapV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/list-flows": {
            "post": {
                "operationId": "listFlowsV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/list-flows",
                        "verbLowerCase": "post"
                    }
                },
                "description": "Responds with a list of the flows on the Corda node.",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ListFlowsV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListFlowsV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/diagnose-node": {
            "post": {
                "operationId": "diagnoseNodeV1",
                "x-hyperledger-cacti": {
                    "http": {
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/diagnose-node",
                        "verbLowerCase": "post"
                    }
                },
                "description": "Responds with diagnostic information about the Corda node",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DiagnoseNodeV1Request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiagnoseNodeV1Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/getFlow/{holdingIDShortHash}": {
            "get": {
                "x-hyperledger-cactus": {
                    "http": {
                        "verbLowerCase": "get",
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/getFlow/{holdingIDShortHash}"
                    }
                },
                "operationId": "getFlowV1",
                "summary": "This method returns an array containing the statuses of all flows running for a specified holding identity. An empty array is returned if there are no flows running.",
                "parameters": [
                    {
                        "name": "holdingIDShortHash",
                        "in": "path",
                        "required": true,
                        "description": "Holding identity short hash",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlowStatusV5Responses"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/flow/{holdingIDShortHash}/{clientRequestID}": {
            "get": {
                "x-hyperledger-cactus": {
                    "http": {
                        "verbLowerCase": "get",
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/flow/{holdingIDShortHash}/{clientRequestID}"
                    }
                },
                "operationId": "flowStatusResponse",
                "summary": "This method gets the current status of the specified flow instance.",
                "parameters": [
                    {
                        "name": "holdingIDShortHash",
                        "in": "path",
                        "required": true,
                        "description": "Holding identity short hash",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "clientRequestID",
                        "in": "path",
                        "required": true,
                        "description": "Client request ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlowStatusV5Response"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/startFlow": {
            "post": {
                "x-hyperledger-cactus": {
                    "http": {
                        "verbLowerCase": "post",
                        "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/startFlow"
                    }
                },
                "operationId": "startFlowV1",
                "summary": "This method starts a new instance for the specified flow for the specified holding identity.",
                "requestBody": {
                    "description": "Request body for starting a flow",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartFlowV5Request"
                            }
                        }
                    }
                },
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StartFlowV1Response"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}